# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: DigiHostel API
  version: 0.1.0
  description: >
    Source contract for the DigiHostel API (ADR-007). Health check plus the
    first two real business workflows, Parent Leave Approval and Student Leave
    Request creation/viewing (docs/leave-approval-workflow.md).
servers:
  - url: /api/v1
paths:
...
  /devices/challenge:
    post:
      operationId: requestDeviceChallenge
      tags: [devices]
      summary: >
        Request a device-registration challenge (ADR-003 implementation)
      security: [{ bearerAuth: [] }]
      description: >
        Server-controlled first step of trusted-device registration: issues a
        single-use, short-lived, cryptographically random nonce bound to the
        caller's own parent identity. The client passes this nonce into the
        platform attestation call (Play Integrity on Android) and submits the
        resulting token to POST /devices/register — the client is never the
        authority for the challenge value itself.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestDeviceChallengeBody"
      responses:
        "200":
          description: A real, server-issued challenge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceChallenge"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "403":
          description: Authenticated caller is not a parent/guardian
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /devices/register:
    post:
      operationId: registerDevice
      tags: [devices]
      summary: >
        Submit platform attestation to complete device registration (ADR-003
        implementation)
      security: [{ bearerAuth: [] }]
      description: >
        Redeems a previously-issued, single-use challenge and verifies the
        submitted attestation token server-side (Google Play Integrity
        Standard API on Android) before ever creating a trusted_devices row.
        The client can never self-declare trust — every outcome other than a
        genuine, server-verified PASS leaves the device untrusted.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDeviceBody"
      responses:
        "200":
          description: Attestation verified — the device is now trusted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrustedDevice"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          description: >
            Unauthenticated, or the challenge is invalid/expired/already
            consumed/belongs to a different caller
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "403":
          description: Attestation was verified but rejected by policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "502":
          description: The attestation provider could not be reached/verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "503":
          description: No attestation provider is configured in this environment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth-issued JWT (ADR-014).
...
  schemas:
...
    LeaveRequestStatus:
      type: string
      description: >
        Full state vocabulary already defined by the schema
        (packages/db/src/schema/enums.ts, leave_request_status) — this
        endpoint set only ever produces approved/rejected transitions today;
        the escalation-notification states are included because they are
        part of the same enum, not because this task creates them.
      enum:
        - pending
        - father_notified
        - mother_notified
        - guardian_notified
        - approved
        - rejected
        - in_app_call
        - manual_verification
        - expired
...
    DevicePlatform:
      type: string
      description: Mirrors packages/db/src/schema/enums.ts's device_platform enum.
      enum: [android, ios]
    RequestDeviceChallengeBody:
      type: object
      required: [platform]
      additionalProperties: false
      properties:
        platform:
          $ref: "#/components/schemas/DevicePlatform"
    DeviceChallenge:
      type: object
      required: [challengeId, nonce, expiresAt]
      properties:
        challengeId:
          type: string
          format: uuid
        nonce:
          type: string
          description: >
            Pass this exact value into the platform attestation call (e.g.
            Play Integrity's requestHash/nonce parameter) so the resulting
            token is cryptographically bound to this specific challenge.
        expiresAt:
          type: string
          format: date-time
    RegisterDeviceBody:
      type: object
      required: [challengeId, platform, attestationToken, deviceFingerprint]
      additionalProperties: false
      properties:
        challengeId:
          type: string
          format: uuid
        platform:
          $ref: "#/components/schemas/DevicePlatform"
        attestationToken:
          type: string
          minLength: 1
          description: >
            The opaque token the platform attestation API produced (a Play
            Integrity integrity token on Android) — never inspected or
            decoded by the client, only relayed to the backend.
        deviceFingerprint:
          type: string
          minLength: 1
          description: >
            This installation's own random, app-generated identifier
            (services/deviceIdentity — never a hardware serial/IMEI/Android
            ID), used only to recognize the same installation on a future
            listTrustedDevices() call, never as attestation proof itself.
    TrustedDevice:
      type: object
      required: [id, platform, registeredAt]
      properties:
        id:
          type: string
          format: uuid
        platform:
          $ref: "#/components/schemas/DevicePlatform"
        registeredAt:
          type: string
          format: date-time
...
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            currentStatus:
              $ref: "#/components/schemas/LeaveRequestStatus"
            requestId:
              type: string
              description: >-
                Server-generated per-request identifier (F-07 production
                observability). Present on unexpected (5xx) and
                framework-rejected (4xx) errors; hand this value to support
                to correlate a client-visible failure with server-side logs.
                Not present on typed domain errors that routes map to a
                response themselves (e.g. leave-request 404/409).
...
  responses:
...
    ValidationError:
      description: Request body/params failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"